refactor: Move data_directory types into /shared
