+++ pubky-homeserver/src/lib.rs
@@ -28,7 +28,17 @@ pub mod tracing;
 pub use admin_server::{AdminServer, AdminServerBuildError};
 pub use app_context::{AppContext, AppContextConversionError};
 pub use client_server::{ClientServer, ClientServerBuildError};
-pub use data_directory::*;
+#[cfg(any(test, feature = "testing"))]
+pub use data_directory::MockDataDir;
+pub use data_directory::{
+    storage_config, AdminToml, ConfigReadError, ConfigToml, DataDir, LoggingToml, MetricsToml,
+    PersistentDataDir,
+};
 pub use homeserver_app::{HomeserverApp, HomeserverAppBuildError};
 pub use metrics_server::{MetricsServer, MetricsServerBuildError};
 pub use persistence::sql::ConnectionString;
+pub use shared::quota::{
+    BandwidthQuota, DefaultQuotasToml, GlobPattern, HttpMethod, LimitKey, LimitKeyType, PathLimit,
+    RequestCountQuota, TimeUnit,
+};
+pub use shared::{Domain, DomainPort, SignupMode};
