+++ pubky-homeserver/src/persistence/files/write_path_layer.rs
@@ -185,7 +185,7 @@ mod tests {
     };
     use crate::persistence::sql::SqlDb;
     use crate::services::user_service::UserService;
-    use crate::shared::user_quota::UserQuota;
+    use crate::shared::quota::UserQuota;
     use crate::shared::webdav::StoragePath;
 
     use super::*;
