+++ pubky-homeserver/src/shared/mod.rs
@@ -1,10 +1,16 @@
+mod domain;
+mod domain_port;
 mod http_error;
 mod pubkey_path_validator;
+pub(crate) mod quota;
+mod signup_mode;
 pub(crate) mod toml_merge;
-pub(crate) mod user_quota;
 mod utils;
 pub(crate) mod webdav;
 
+pub use domain::Domain;
+pub use domain_port::DomainPort;
 pub(crate) use http_error::{HttpError, HttpResult};
 pub(crate) use pubkey_path_validator::Z32Pubkey;
+pub use signup_mode::SignupMode;
 pub(crate) use utils::{parse_bool, timestamp_to_sqlx_datetime};
