+++ pubky-homeserver/src/shared/quota/mod.rs
@@ -1,4 +1,5 @@
 mod bandwidth_quota;
+mod default_quotas;
 mod glob_pattern;
 mod http_method;
 mod limit_key;
@@ -6,11 +7,14 @@ mod path_limit;
 pub(crate) mod rate_unit;
 mod request_count_quota;
 mod time_unit;
+pub(crate) mod user_quota;
 
 pub use bandwidth_quota::BandwidthQuota;
+pub use default_quotas::DefaultQuotasToml;
 pub use glob_pattern::GlobPattern;
 pub use http_method::HttpMethod;
 pub use limit_key::{LimitKey, LimitKeyType};
-pub use path_limit::*;
+pub use path_limit::PathLimit;
 pub use request_count_quota::RequestCountQuota;
 pub use time_unit::TimeUnit;
+pub(crate) use user_quota::{UserQuota, UserQuotaPatch};
