+++ pubky-homeserver/src/shared/signup_mode.rs
